Decouple from Next
